requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
decorator==5.1.1
distro==1.7.0
distro-info===1.1build1
Django==4.1.1
django-extensions==3.2.1
django_debug_toolbar==3.8.1
executing==1.1.0
Faker==15.0.0
flake8==5.0.4
flake8-builtins==2.0.0
flake8-import-order==0.18.1
flake8-print==5.0.0
httplib2==0.20.2
importlib-metadata==4.6.4
ipython==8.5.0
jedi==0.18.1
jeepney==0.7.1
keyring==23.5.0
kombu==5.2.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
systemd-python==234
traitlets==5.4.0
tzdata==2022.4
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
zipp==1.0.0
